Added nested array support
CRUD operations for arrays which are direct children of an array. Modified Path initlisation to take that into account. Updated People files to use them in test/playground.
