+++ Playground/People.plist
@@ -28,6 +28,20 @@
 				<string>party</string>
 				<string>tennis</string>
 			</array>
+			<key>running_records</key>
+			<array>
+				<array>
+					<integer>10</integer>
+					<integer>12</integer>
+					<integer>9</integer>
+					<integer>10</integer>
+				</array>
+				<array>
+					<integer>9</integer>
+					<integer>12</integer>
+					<integer>11</integer>
+				</array>
+			</array>
 		</dict>
 		<key>Suzanne</key>
 		<dict>
