+++ Tests/ScoutTests/Integration/PathExplorerTests_Integration.swift
@@ -1,5 +1,5 @@
 import XCTest
-@testable import Scout
+import Scout
 
 final class PathExplorerTestsIntegration: XCTestCase {
 
@@ -12,6 +12,7 @@ final class PathExplorerTestsIntegration: XCTestCase {
     let tomLastHobby: Path = ["people", "Tom", "hobbies", -1]
     let suzanneFirstMovieTitle: Path = ["people", "Suzanne", "movies", 0, "title"]
     let suzanneLastMovieTitle: Path = ["people", "Suzanne", "movies", -1, "title"]
+    let robertRunningRecordsSecondFirst: Path = ["people", "Robert", "running_records", 1, 0]
 
     // MARK: - Properties
 
@@ -60,35 +61,8 @@ final class PathExplorerTestsIntegration: XCTestCase {
         testPathExplorersGet(path: suzanneLastMovieTitle, value: "What about today?")
     }
 
-    // -- Helpers
-
-    func testGet<Explorer: PathExplorer, Value: KeyAllowedType>(path: Path, explorer: Explorer, value: Value, file: StaticString = #file, line: UInt = #line) {
-        do {
-            switch value {
-            case let stringValue as String:
-                let value = try explorer.get(path).string
-                XCTAssertEqual(value, stringValue, "\(explorer.format): \(value?.description ?? "nil") is not equal to \(stringValue)", file: file, line: line)
-            case let intValue as Int:
-                let value = try explorer.get(path).int
-                XCTAssertEqual(value, intValue, "\(explorer.format): \(value?.description ?? "nil") is not equal to \(intValue)", file: file, line: line)
-            case let doubleValue as Double:
-                let value = try explorer.get(path).real
-                XCTAssertEqual(value, doubleValue, "\(explorer.format): \(value?.description ?? "nil") is not equal to \(doubleValue)", file: file, line: line)
-            case let boolValue as Bool:
-                let value = try explorer.get(path).bool
-                XCTAssertEqual(value, boolValue, "\(explorer.format): \(value?.description ?? "nil") is not equal to \(boolValue)", file: file, line: line)
-            default:
-                assertionFailure("Value not a KeyAllowedType")
-            }
-        } catch {
-            XCTFail("\(explorer.format): \(error.localizedDescription)", file: file, line: line)
-        }
-    }
-
-    func testPathExplorersGet<Value: KeyAllowedType>(path: Path, value: Value, file: StaticString = #file, line: UInt = #line) {
-        testGet(path: path, explorer: json, value: value, file: file, line: line)
-        testGet(path: path, explorer: plist, value: value, file: file, line: line)
-        testGet(path: path, explorer: xml, value: value, file: file, line: line)
+    func testGetNestedArrayInArrayIndex() {
+        testPathExplorersGet(path: robertRunningRecordsSecondFirst, value: 9)
     }
 
     // MARK: Set
@@ -113,27 +87,11 @@ final class PathExplorerTestsIntegration: XCTestCase {
         testPathExplorersSet(path: suzanneLastMovieTitle, value: "Never gonna die")
     }
 
-    // -- Helpers
-
-    func testSet<Explorer: PathExplorer, Value: KeyAllowedType>(path: Path, explorer: Explorer, value: Value, file: StaticString = #file, line: UInt = #line) {
-        var explorer = explorer
-        do {
-            try explorer.set(path, to: value)
-        } catch {
-            XCTFail("\(explorer.format): \(error.localizedDescription)", file: file, line: line)
-            return
-        }
-
-        testGet(path: path, explorer: explorer, value: value, file: file, line: line)
+    func testSetNestedArrayInArrayIndex() {
+        testPathExplorersSet(path: robertRunningRecordsSecondFirst, value: 25)
     }
 
-    func testPathExplorersSet<Value: KeyAllowedType>(path: Path, value: Value, file: StaticString = #file, line: UInt = #line) {
-        testSet(path: path, explorer: json, value: value, file: file, line: line)
-        testSet(path: path, explorer: plist, value: value, file: file, line: line)
-        testSet(path: path, explorer: xml, value: value, file: file, line: line)
-    }
-
-    // MARK: - Delete
+    // MARK: Delete
 
     func testDeleteDictKey() {
         testPathExplorersDelete(path: tomHeight)
@@ -163,37 +121,8 @@ final class PathExplorerTestsIntegration: XCTestCase {
         testPathExplorersDelete(path: tomHobbies)
     }
 
-    // -- Helpers
-
-    func testDelete<Explorer: PathExplorer>(path: Path, explorer: Explorer, file: StaticString = #file, line: UInt = #line) {
-        var explorer = explorer
-        let value: String
-        do {
-            value = try explorer.get(path).stringValue
-            try explorer.delete(path)
-        } catch {
-            XCTFail("\(explorer.format): \(error.localizedDescription)", file: file, line: line)
-            return
-        }
-
-        do {
-            let newValue = try explorer.get(path).stringValue
-            // if the value was deleted at -1 in an array, the array might still has a last value. So make sure it's not the same
-            XCTAssertNotEqual(value, newValue, "\(explorer.format): The value at \(path) has not been deleted", file: file, line: line)
-        } catch {
-            switch error {
-            case PathExplorerError.subscriptWrongIndex, PathExplorerError.subscriptMissingKey:
-                return
-            default:
-                XCTFail("Wrong error for getting a non-exiting value: \(error). Should be 'subscriptWrongIndex' or 'subscriptMissingKey'")
-            }
-        }
-    }
-
-    func testPathExplorersDelete(path: Path, file: StaticString = #file, line: UInt = #line) {
-        testDelete(path: path, explorer: json, file: file, line: line)
-        testDelete(path: path, explorer: plist, file: file, line: line)
-        testDelete(path: path, explorer: xml, file: file, line: line)
+    func testDeleteNestedArrayInArrayIndex() {
+        testPathExplorersDelete(path: robertRunningRecordsSecondFirst)
     }
 
     // MARK: Add
@@ -211,11 +140,12 @@ final class PathExplorerTestsIntegration: XCTestCase {
     }
 
     func testAddNestedArrayIndex() {
-        testPathExplorersSet(path: suzanneFirstMovieTitle, value: "Never gonna die")
+        // adding value to an existing key should act like setting it
+        testPathExplorersAdd(path: suzanneFirstMovieTitle, value: "Never gonna die")
     }
 
     func testAddNestedArrayLastIndex() {
-        testPathExplorersSet(path: suzanneLastMovieTitle, value: "Never gonna die")
+        testPathExplorersAdd(path: suzanneLastMovieTitle, value: "Never gonna die")
     }
 
     func testAddNewDict() {
@@ -234,23 +164,12 @@ final class PathExplorerTestsIntegration: XCTestCase {
         testPathExplorersAdd(path: ["people", "Cecil", "secretLoves", 0], value: "Candies")
     }
 
-    // -- Helpers
-
-    func testAdd<Explorer: PathExplorer, Value: KeyAllowedType>(path: Path, explorer: Explorer, value: Value, file: StaticString = #file, line: UInt = #line) {
-        var explorer = explorer
-        do {
-            try explorer.add(value, at: path)
-        } catch {
-            XCTFail("\(explorer.format): \(error.localizedDescription)", file: file, line: line)
-            return
-        }
-
-        testGet(path: path, explorer: explorer, value: value, file: file, line: line)
+    func testAddNestedArrayInArray() {
+        testPathExplorersAdd(path: ["people", "Robert", "running_records", 0, -1], value: 30)
     }
 
-    func testPathExplorersAdd<Value: KeyAllowedType>(path: Path, value: Value, file: StaticString = #file, line: UInt = #line) {
-        testAdd(path: path, explorer: json, value: value, file: file, line: line)
-        testAdd(path: path, explorer: plist, value: value, file: file, line: line)
-        testAdd(path: path, explorer: xml, value: value, file: file, line: line)
+    func testAddNestedArrayTwoLevels() {
+        // I know... who could possibly use that, right?
+        testPathExplorersAdd(path: ["people", "Robert", "running_records", 0, -1, 0], value: 25)
     }
 }
